Refactor folder structure and update link paths
- Restructured folder organization
- Updated login redirection path to /user/verify-email
- Changed email verification URL in the sendVerificationEmailService
